Test fixes!

+++ test.js
@@ -1,6 +1,9 @@
 var exec = require('child_process').exec;
 var test = require('tape');
 
+// Travis-CI doesn't have node in its path, but we can get it from npm test
+var node = process.argv[2];
+
 function verifyOutput(t, stdout, stderr, firstError, lastError) {
   t.equal(stdout, '', 'should not output anything to stdout');
 
@@ -49,45 +52,48 @@ function verifyError(t, err, code, stdout, stderr, firstError, lastError) {
 }
 
 test('Nothing output when no errors', function(t) {
-  exec('node tester.js', {env: {TEST_NUMBER: 1}}, function(err, stdout, stderr) {
+  exec(node + ' tester.js', {env: {TEST_NUMBER: 1}}, function(err, stdout, stderr) {
     verifyNoError(t, err, stdout, stderr);
   });
 });
 
 test('Output one error', function(t) {
-  exec('node tester.js', {env: {TEST_NUMBER: 2}}, function(err, stdout, stderr) {
+  exec(node +' tester.js', {env: {TEST_NUMBER: 2}}, function(err, stdout, stderr) {
     verifyNoError(t, err, stdout, stderr, 1, 1);
   });
 });
 
 test('Output two errors', function(t) {
-  exec('node tester.js', {env: {TEST_NUMBER: 3}}, function(err, stdout, stderr) {
+  exec(node + ' tester.js', {env: {TEST_NUMBER: 3}}, function(err, stdout, stderr) {
     verifyNoError(t, err, stdout, stderr, 1, 2);
   });
 });
 
 test('Output last ten errors', function(t) {
-  exec('node tester.js', {env: {TEST_NUMBER: 4}}, function(err, stdout, stderr) {
+  exec(node + ' tester.js', {env: {TEST_NUMBER: 4}}, function(err, stdout, stderr) {
     verifyNoError(t, err, stdout, stderr, 41, 50);
   });
 });
 
-test('Output five errors after killed by SIGINT', function(t) {
-  child = exec('node tester.js', {env: {TEST_NUMBER: 5}}, function(err, stdout, stderr) {
-    verifyError(t, err, 130, stdout, stderr, 1, 5);
+// Sadly, Travis-CI won't let us kill child processes
+if (!process.env.TRAVIS) {
+  test('Output five errors after killed by SIGINT', function(t) {
+    child = exec(node + ' tester.js', {env: {TEST_NUMBER: 5}}, function(err, stdout, stderr) {
+      verifyError(t, err, 130, stdout, stderr, 1, 5);
+    });
+
+    setTimeout(function() {
+      child.kill('SIGINT');
+    }, 1000);
   });
 
-  setTimeout(function() {
-    child.kill('SIGINT');
-  }, 1000);
-});
+  test('Output five errors after killed by SIGTERM', function(t) {
+    child = exec(node + ' tester.js', {env: {TEST_NUMBER: 5}}, function(err, stdout, stderr) {
+      verifyError(t, err, 143, stdout, stderr, 1, 5);
+    });
 
-test('Output five errors after killed by SIGTERM', function(t) {
-  child = exec('node tester.js', {env: {TEST_NUMBER: 5}}, function(err, stdout, stderr) {
-    verifyError(t, err, 143, stdout, stderr, 1, 5);
+    setTimeout(function() {
+      child.kill('SIGTERM');
+    }, 1000);
   });
-
-  setTimeout(function() {
-    child.kill('SIGTERM');
-  }, 1000);
-});
+}
